Light refactoring and cleanup
Removed duplicated call to create database, and generally tidied up.

+++ app/src/main/java/ca/cplyon/cointime/ui/main/CoinListAdapter.kt
@@ -82,7 +82,7 @@ class CoinListAdapter internal constructor(
     }
 
     override fun getItemId(position: Int): Long {
-        return coins[position].coinId.toLong()
+        return coins[position].coinId
     }
 
     override fun getFilter(): Filter {
